Flashlight Almost Dead Indicator
There is now an indicator for when the flashlight is going to die, and
a battery that goes with the flashlight that can be added/removed from
a flashlight. (TODO: Dealing with multiple batteries, although this
could be focused after the “demo.”) Next step before push to master:
Better Documentation on added stuff?